+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osh/OSHEntityFilter.java
@@ -1,0 +1,8 @@
+package org.heigit.ohsome.oshdb.osh;
+
+import java.io.Serializable;
+import java.util.function.Predicate;
+
+public interface OSHEntityFilter extends Predicate<OSHEntity>, Serializable {
+
+}