+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osm/OSMEntityFilter.java
@@ -1,0 +1,8 @@
+package org.heigit.ohsome.oshdb.osm;
+
+import java.io.Serializable;
+import java.util.function.Predicate;
+
+public interface OSMEntityFilter extends Predicate<OSMEntity>, Serializable {
+
+}